--- scheduling/scheduling-db-scheduler/src/main/java/io/koraframework/scheduling/db/annotation/ScheduleWithCron.java
@@ -1,16 +1,0 @@
-package io.koraframework.scheduling.db.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.CLASS)
-public @interface ScheduleWithCron {
-    String value() default "";
-
-    String name() default "";
-
-    String config() default "";
-}
