--- scheduling/scheduling-db-scheduler/src/main/java/io/koraframework/scheduling/db/job/SchedulingDbJob.java
@@ -1,8 +1,0 @@
-package io.koraframework.scheduling.db.job;
-
-import com.github.kagkarlsson.scheduler.task.Task;
-
-public interface SchedulingDbJob {
-
-    Task<?> task();
-}
