+++ scheduling/scheduling-db-scheduler/src/main/java/io/koraframework/scheduling/db/scheduler/job/CronJob.java
@@ -1,4 +1,4 @@
-package io.koraframework.scheduling.db.job;
+package io.koraframework.scheduling.db.scheduler.job;
 
 import com.github.kagkarlsson.scheduler.task.helper.RecurringTask;
 import com.github.kagkarlsson.scheduler.task.helper.Tasks;
